requirements.txt
Pillow
tqdm
customtkinter
